changed from int to double close #92

+++ src/main/java/il/org/spartan/utils/Point.java
@@ -5,10 +5,10 @@ import java.util.*;
 import fluent.ly.*;
 
 public class Point {
-  int x;
-  int y;
+  double x;
+  double y;
 
-  public Point(int x, int y) {
+  public Point(double x, double y) {
     this.x = x;
     this.y = y;
   }
